Updated tab linking.
Linking to specific tabs is now done using the non-localized version of the tab name (i.e. the #hash is a string defined in i18n.js). However, it still falls back to the actual tab name. Additionally, the URL no longer changes when switching tabs; in Chrome and Safari extensions the URL is hidden anyway.